requirements.txt
streamlit==1.40.1
Pillow==9.5.0
